Remove Shipyard-specific contributing docs
These docs are so detailed, they are very difficult to keep up-to-date
and are not much quicker to understand than just looking at the code.
Given the need for the user to additionally check the code and
differentiate old docs from reality, having the docs creates more work
than they save.

Additionally, the building and testing targets and options are becoming
uniform across all Submariner projects, and therefore are better covered
only once in the general Building and Testing docs.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>