Doc that "iptables" should always be lowercase
The iptables application consistently uses all-lowercase as their proper
noun and therefore trademark. It seems we should follow their trademark,
but should do our best to avoid needing to start a sentence with
"iptables" and cause an inconsistency with other grammar norms. There
does not seem to be consensus about capitalizing words like this at the
beginning of sentences.

A summary of guidance to support this rule follows.

There does not seem to be any guidance for third-party application names
in the K8s Docs Style Guide:

https://kubernetes.io/docs/contribute/style/style-guide/

In the Fedora Manual for Technical Writers, there is guidance for
application names:

"When used as a proper name, use the capitalization of the product. When
used as a command, use lowercase as appropriate, such as ‘To start GCC,
type gcc’.
Note: ‘vi’ is always lowercase."

All-lowercase "iptables" is used elsewhere in that doc:

"For instructions on how to construct iptables rules... Less confusing"

There is also guidance for avoiding using always-lowercase words at the
beginning of sentences:

"hostname Correct. One word, lowercase. Capitalize when used at the
beginning of a sentence, but try to reword the sentence to avoid this."

https://rlandmann.fedorapeople.org/ossg/html-single/

On English StackExchange, an answer about trademark capitalization seems
to support following the "innovations in capitalization and punctuation
invented by trademark owners" but avoiding starting sentences with such
words when possible.

https://english.stackexchange.com/a/2942/159794

In the iptables manual, it is capitalized as the first word in the page
but not capitalized elsewhere, including at the start of subsequent
paragraphs and sentences.

https://linux.die.net/man/8/iptables

On Wikipedia, it's always lowercase, even at the start of sentences:

https://en.wikipedia.org/wiki/Iptables

Similar tools are also always lowercase on Wikipedia:

https://en.wikipedia.org/wiki/Nftables
https://en.wikipedia.org/wiki/Arptables

In the ArchLinux wiki, it's lowercase even at the start of sentences:

https://wiki.archlinux.org/index.php/Iptables

However, in the CentOS wiki, it's uppercase at the start of sentences:

https://wiki.centos.org/HowTos/Network/IPTables

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>