Doc that "kind" is always lowercase
The kind project is very consistent in using lowercase throughout their
docs, including at the start of paragraphs and sentences.

https://kind.sigs.k8s.io/

Submariner should avoid starting a sentence with an always-lowercase
word to mitigate conflicts with other grammar norms.

The K8s Docs Docs don't provide guidance here, but see the iptables
commit message for other guidance and norms.

For example, in the Fedora Technical Writer's Guide, there's guidance
for avoiding using always-lowercase words at the beginning of sentences:

"hostname Correct. One word, lowercase. Capitalize when used at the
beginning of a sentence, but try to reword the sentence to avoid this."

https://rlandmann.fedorapeople.org/ossg/html-sing

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>