Remove security.txt path not served by website
In the PR that added security.txt files, it was noted that they were
added to both well-known paths because it didn't seem our website
hosting was serving the one in the hidden .well-known directory.

It turns out the production version of the website also doesn't serve up
that file. Remove it, leaving only the security.txt file at root.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>