Relicense docs to CC BY 4.0 per CNCF requirements
Change the license of the docs from Apache 2 to Creative Commons
Attribution 4.0 International.

This is required by the CNCF as a prerequisite for submitting a project:

"All documentation will be received and made available by the CNCF under
the Creative Commons Attribution 4.0 International License."

https://github.com/cncf/foundation/blob/master/charter.md#11-ip-policy

License text copied from:

https://raw.githubusercontent.com/kubernetes/website/
383e40f978334792e7ca30399845c9bf38ea813c/LICENSE

Approvals for relicensing will take place on the PR for this commit.

Closes: #438
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>