Remove numbering of external resources, date-order (#527)
To ensure there's no bias to the "recommended" ordering of external
resources, use bulleted lists instead of numbered lists and order by
date.

Add dates to links for clarity when making updates. Dates are rounded to
months, but are ordered at day-level when there is a tie at month-level.

Add the date inside the link, not after it, to avoid ugly wrapping that
happens when long link-text is followed by short non-link-text. This
issue still happens with the slide links, but the problem is minimized.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>