Update getting started deployment docs
Update the deployment summary in the getting started docs to reflect the
new reality that the Operator is always used, either by subctl, Helm, or
directly. Mention new subctl features as one reason it is recommended.

This section was part of the UX failure of a recent IBM Twitch stream.

https://www.youtube.com/watch?v=98rC2OOW20w

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>