Fix Go-based subctl install docs
Because we have `replace` statements in `submariner-operator/go.mod` to
pin the K8s version, when run from outside of the submariner-operator
repo `go install` declines to build `subctl` with:

The go.mod file for the module providing named packages contains one or
more replace directives. It must not contain directives that would cause
it to be interpreted differently than if it were the main module.

This issue is described upstream here:

https://github.com/golang/go/issues/44840#issuecomment-967294447

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>