Update "Deploy Automatically" for 'make deploy'
The "Deploy Automatically" directions use 'make deploy' then give an
example on how to test the deployment. By default, lighthouse was not
being deployed, so test example was failing. Update to
'make deploy using=lighthouse'.

In debugging this, took a while to stumble on how to customize a Submariner
deployment. So also adding a shortcut to the existing documentation on
customizing to this section in the document. The customization example
failed with syntax error, so updating this section as well to the current
syntax.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>