Use inline HTML for heading tags, not Gitiles MD
This syntax for creating ID tags for headings (typically with complex
names) isn't supported in CommonMark, which is the standard used by our
MD linter. It's valid syntax in Gitiles MD, which is why the links work.

Upstream feature request to add support for this type of header ID:
DavidAnson/markdownlint#570

This is required for submariner-io/submariner-website#827.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>